checking existing file for data

+++ gateio_main.py
@@ -1,6 +1,7 @@
 import asyncio
+import json
 import time
-
+import os
 import requests
 import pandas as pd
 from shared import *
@@ -16,12 +17,18 @@ def common_markets_filtering():
 
 
 if __name__ == '__main__':
-    spot_market_list = spot_market.spot_markets_list()
-    spot_market_list = [market for market in spot_market_list if market['trade_status'] == 'tradable']
-    spot_quote_market = spot_market.spot_quote_tradable_markets(spot_market_list)
-    # returns df with combinations of all usdt and remove unwanted rows from df
-    arb_df = arbitrage_handle.create_quote_df(spot_quote_market)
-    arb_df.to_json('arb_df_bak.json', orient='records')  # Overwrite existing file if exists
+    file_exists = os.path.exists('arb_df_bak.json')
+    if file_exists:
+        with open('arb_df_bak.json', 'r') as file:
+            arb_df = json.load(file)
+    else:
+        spot_market_list = spot_market.spot_markets_list()
+        spot_market_list = [market for market in spot_market_list if market['trade_status'] == 'tradable']
+        spot_quote_market = spot_market.spot_quote_tradable_markets(spot_market_list)
+        # returns df with combinations of all usdt and remove unwanted rows from df
+        arb_df = arbitrage_handle.create_quote_df(spot_quote_market)
+        arb_df.to_json('arb_df_bak.json', orient='records')  # Overwrite existing file if exists
+
     while True:
         print("\033[H\033[J")  # clear terminal
         print(arb_df)
